+++ lib/src/command/enum/enum_write_file.dart
@@ -295,11 +295,11 @@ $maybeMapNullValueSb2
   int compareTo($nameClass other) => index.compareTo(other.index);
 
  
-  @override
-  String toString() => '$nameClass.\$$nameValueFirst';
 
 ''');
 
+  // @override
+  // String toString() => '$nameClass.\$$nameValueFirst';
 /* ****************************** */
 /* ******************************* */